+++ resources/vertex_framed.glsl
@@ -2,10 +2,10 @@
 
 in layout(location = 0) vec2 a_corner;
 in layout(location = 1) vec2 a_size;
-in layout(location = 2) float a_layer;
+in layout(location = 2) int  a_layer;
 out flat vec2 vg_corner;
 out flat vec2 vg_size;
-out flat float vg_layer;
+out flat int  vg_layer;
 void main()
 {
     vg_corner = a_corner;
